requirements.txt
openai
langchain
langchain-community
langchain-openai
langchain-core
Jinja2
jinja2-time
flask
pep8
python-dateutil
python-docx
python-editor
python-lsp-black
python-lsp-jsonrpc
python-lsp-server
pytz
PyPDF2
faiss-cpu
pypdf
python-dotenv